+++ app/src/main/kotlin/com/simplemobiletools/smsmessenger/adapters/ContactsAdapter.kt
@@ -67,7 +67,7 @@ class ContactsAdapter(activity: SimpleActivity, var contacts: ArrayList<SimpleCo
             }
 
             findViewById<TextView>(R.id.item_contact_number).apply {
-                text = contact.phoneNumber
+                text = contact.phoneNumbers.first()
                 setTextColor(textColor)
                 setTextSize(TypedValue.COMPLEX_UNIT_PX, fontSize)
             }