+++ app/src/main/kotlin/com/simplemobiletools/smsmessenger/adapters/ThreadAdapter.kt
@@ -29,17 +29,17 @@ import com.simplemobiletools.commons.views.FastScroller
 import com.simplemobiletools.commons.views.MyRecyclerView
 import com.simplemobiletools.smsmessenger.R
 import com.simplemobiletools.smsmessenger.activities.SimpleActivity
+import com.simplemobiletools.smsmessenger.dialogs.SelectTextDialog
 import com.simplemobiletools.smsmessenger.extensions.deleteMessage
 import com.simplemobiletools.smsmessenger.helpers.*
-import com.simplemobiletools.smsmessenger.models.Message
-import com.simplemobiletools.smsmessenger.models.ThreadDateTime
-import com.simplemobiletools.smsmessenger.models.ThreadError
-import com.simplemobiletools.smsmessenger.models.ThreadItem
+import com.simplemobiletools.smsmessenger.models.*
 import kotlinx.android.synthetic.main.item_attachment_image.view.*
 import kotlinx.android.synthetic.main.item_received_message.view.*
 import kotlinx.android.synthetic.main.item_received_unknown_attachment.view.*
 import kotlinx.android.synthetic.main.item_sent_unknown_attachment.view.*
 import kotlinx.android.synthetic.main.item_thread_date_time.view.*
+import kotlinx.android.synthetic.main.item_thread_error.view.*
+import kotlinx.android.synthetic.main.item_thread_success.view.*
 
 class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem>, recyclerView: MyRecyclerView, fastScroller: FastScroller,
                     itemClick: (Any) -> Unit) : MyRecyclerViewAdapter(activity, recyclerView, fastScroller, itemClick) {
@@ -60,6 +60,7 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
         menu.apply {
             findItem(R.id.cab_copy_to_clipboard).isVisible = isOneItemSelected
             findItem(R.id.cab_share).isVisible = isOneItemSelected
+            findItem(R.id.cab_select_text).isVisible = isOneItemSelected
         }
     }
 
@@ -71,7 +72,7 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
         when (id) {
             R.id.cab_copy_to_clipboard -> copyToClipboard()
             R.id.cab_share -> shareText()
-            R.id.cab_select_all -> selectAll()
+            R.id.cab_select_text -> selectText()
             R.id.cab_delete -> askConfirmDelete()
         }
     }
@@ -93,6 +94,7 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
             THREAD_DATE_TIME -> R.layout.item_thread_date_time
             THREAD_RECEIVED_MESSAGE -> R.layout.item_received_message
             THREAD_SENT_MESSAGE_ERROR -> R.layout.item_thread_error
+            THREAD_SENT_MESSAGE_SUCCESS -> R.layout.item_thread_success
             else -> R.layout.item_sent_message
         }
         return createViewHolder(layout, parent)
@@ -101,10 +103,11 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = messages[position]
         holder.bindView(item, true, item is Message) { itemView, layoutPosition ->
-            if (item is ThreadDateTime) {
-                setupDateTime(itemView, item)
-            } else if (item !is ThreadError) {
-                setupView(itemView, item as Message)
+            when (item) {
+                is ThreadDateTime -> setupDateTime(itemView, item)
+                is ThreadSuccess -> setupThreadSuccess(itemView)
+                is ThreadError -> setupThreadError(itemView)
+                else -> setupView(itemView, item as Message)
             }
         }
         bindViewHolder(holder)
@@ -118,6 +121,7 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
             item is ThreadDateTime -> THREAD_DATE_TIME
             (messages[position] as? Message)?.isReceivedMessage() == true -> THREAD_RECEIVED_MESSAGE
             item is ThreadError -> THREAD_SENT_MESSAGE_ERROR
+            item is ThreadSuccess -> THREAD_SENT_MESSAGE_SUCCESS
             else -> THREAD_SENT_MESSAGE
         }
     }
@@ -130,6 +134,13 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
     private fun shareText() {
         val firstItem = getSelectedItems().first() as? Message ?: return
         activity.shareTextIntent(firstItem.body)
+    }
+
+    private fun selectText() {
+        val firstItem = getSelectedItems().first() as? Message ?: return
+        if (firstItem.body.trim().isNotEmpty()) {
+            SelectTextDialog(activity, firstItem.body)
+        }
     }
 
     private fun askConfirmDelete() {
@@ -308,5 +319,13 @@ class ThreadAdapter(activity: SimpleActivity, var messages: ArrayList<ThreadItem
                 thread_sim_icon.applyColorFilter(textColor)
             }
         }
+    }
+
+    private fun setupThreadSuccess(view: View) {
+        view.thread_success.applyColorFilter(textColor)
+    }
+
+    private fun setupThreadError(view: View) {
+        view.thread_error.setTextSize(TypedValue.COMPLEX_UNIT_PX, fontSize)
     }
 }