+++ app/src/main/kotlin/com/simplemobiletools/smsmessenger/helpers/Constants.kt
@@ -11,6 +11,7 @@ const val THREAD_ATTACHMENT_URI = "thread_attachment_uri"
 const val THREAD_ATTACHMENT_URIS = "thread_attachment_uris"
 const val USE_SIM_ID_PREFIX = "use_sim_id_"
 const val NOTIFICATION_CHANNEL = "simple_sms_messenger"
+const val SHOW_CHARACTER_COUNTER = "show_character_counter"
 
 private const val PATH = "com.simplemobiletools.smsmessenger.action."
 const val MARK_AS_READ = PATH + "mark_as_read"
@@ -21,6 +22,7 @@ const val THREAD_DATE_TIME = 1
 const val THREAD_RECEIVED_MESSAGE = 2
 const val THREAD_SENT_MESSAGE = 3
 const val THREAD_SENT_MESSAGE_ERROR = 4
+const val THREAD_SENT_MESSAGE_SUCCESS = 5
 
 fun refreshMessages() {
     EventBus.getDefault().post(Events.RefreshMessages())