+++ app/src/main/kotlin/com/simplemobiletools/smsmessenger/receivers/MarkAsReadReceiver.kt
@@ -7,6 +7,7 @@ import com.simplemobiletools.commons.extensions.notificationManager
 import com.simplemobiletools.commons.helpers.ensureBackgroundThread
 import com.simplemobiletools.smsmessenger.extensions.conversationsDB
 import com.simplemobiletools.smsmessenger.extensions.markThreadMessagesRead
+import com.simplemobiletools.smsmessenger.extensions.updateUnreadCountBadge
 import com.simplemobiletools.smsmessenger.helpers.MARK_AS_READ
 import com.simplemobiletools.smsmessenger.helpers.THREAD_ID
 
@@ -19,6 +20,7 @@ class MarkAsReadReceiver : BroadcastReceiver() {
                 ensureBackgroundThread {
                     context.markThreadMessagesRead(threadId)
                     context.conversationsDB.markRead(threadId.toLong())
+                    context.updateUnreadCountBadge(context.conversationsDB.getUnreadConversations())
                 }
             }
         }